+++ packages/web/src/components/gcds-error-message/stories/gcds-error-message.stories.tsx
@@ -56,7 +56,7 @@ const TemplatePlayground = (args) => (`
 export const Default = Template.bind({});
 Default.args = {
   messageId: 'message-id',
-  message: 'This is an error message.',
+  message: 'Error message or validation message.',
 };
 
 // ------ Error message events & props ------
@@ -64,7 +64,7 @@ Default.args = {
 export const Props = Template.bind({});
 Props.args = {
   messageId: 'message-id',
-  message: 'This is an error message.',
+  message: 'Error message or validation message.',
 };
 
 // ------ Error message playground ------
@@ -72,5 +72,5 @@ Props.args = {
 export const Playground = TemplatePlayground.bind({});
 Playground.args = {
   messageId: '',
-  message: 'This is an error message.',
+  message: 'Error message or validation message.',
 };