+++ packages/web/src/components/gcds-radio/stories/gcds-radio.stories.tsx
@@ -135,9 +135,9 @@ const Template = (args) => (`
 export const Default = Template.bind({});
 Default.args = {
   radioId: 'radio',
-  label: 'Radio label',
+  label: 'Label',
   name: 'radio',
-  hint: '',
+  hint: 'Description or example to make the option clearer.',
   required: false,
   disabled: false,
   value: '',